#25 feat: getBooksByAuthorId getAuthorByBookId
toAuthorResponses
toBookResponses

+++ store/src/main/java/store/novabook/store/book/entity/Book.java
@@ -1,6 +1,5 @@
 package store.novabook.store.book.entity;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Entity;
@@ -13,7 +12,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -57,7 +55,7 @@ public class Book {
 	int inventory;
 
 	@NotNull
-	private BigDecimal price;
+	private long price;
 
 	@NotNull
 	boolean isPackaged;
